Add TCPComm for Pi and EV3
This adds the TCPComm Python file for upload to the EV3 and the
Pi. It also adds the ability to send a message from the EV3 to the
pi - they forgot the add the functionality! To add this on line 372
`self.sock.sendall(msg.encode())` was added.

TCPComm is from https://github.com/Ignat-Georgiev/tcpcom and is
licenced under a MIT License