Add Networking to Line Following
This adds basic networking for the line following class, when
created the class will start a connection with the Pi and then await
messages - if a GOTO$X$Y (with X being the x coord and Y being the y
coord) is recived it will call the function to move there. It also
allows for a ARRIVED msg to be sent to the Pi to allow for
authentiction. Finally if STOP is sent the robot SHOULD stop - but
due to the magic of threading it does not

TO-DO Make STOP command stop the robot - probably using a thread.