Add State & Message Support
These changes add support for states for the robot - the Pi is now
able to interprate what a server sends to it and make decisons about
what to get the EV3 to do.
For example it will now only send the robot
to the desination after the parcel has been placed inside it and closed
the lid.
Also changes the server to be the demo room DiCE machine.